+++ src/main/java/org/codinjutsu/tools/jenkins/logic/JenkinsBackgroundLoadingOption.java
@@ -1,0 +1,13 @@
+package org.codinjutsu.tools.jenkins.logic;
+
+import com.intellij.openapi.progress.PerformInBackgroundOption;
+
+public class JenkinsBackgroundLoadingOption implements PerformInBackgroundOption {
+
+    public static final JenkinsBackgroundLoadingOption INSTANCE = new JenkinsBackgroundLoadingOption();
+
+    public boolean shouldStartInBackground() {
+        return true;
+    }
+
+}